+++ CHANGELOG.md
@@ -8,6 +8,7 @@ dev
   maximal length of notifications.
 * Added a new option: `escape_html`. It escapes the `<`, `>`, and `&` HTML
   characters in notification messages.
+* Fix notifications which do not have a timeout, icon, or urgency set.
 
 0.1 (2015-06-21)
 ----------------